requirements.txt
Django
psycopg2-binary
pandapower
matplotlib
sklearn
tqdm
IPython
plotly
django-cors-headers
djangorestframework

